Move new WorkerLocal implementation to the old location
